requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
ffmpeg==1.4
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
numpy==2.1.3
pillow==11.0.0
psutil==6.1.0
python-telegram-bot==21.7
requests==2.32.3
setuptools==75.6.0
sniffio==1.3.1
urllib3==2.2.3
youtube-dl==2021.12.17
yt-dlp==2024.11.18
